requirements.txt
apache-beam==2.31.0
avro-python3==1.9.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.4
fasteners==0.16.3
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-apitools==0.5.31
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==2.23.2
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.2
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-profiler==3.0.5
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-storage==1.41.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-resumable-media==1.3.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
grpcio-gcp==0.2.2
hdfs==2.6.0
httplib2==0.19.1
idna==3.2
numpy==1.20.3
oauth2client==4.1.3
packaging==21.0
pandas==1.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
